+++ oauth2_http/java/com/google/auth/oauth2/StsRequestHandler.java
@@ -50,7 +50,7 @@ import java.util.List;
 import javax.annotation.Nullable;
 
 /** Implements the OAuth 2.0 token exchange based on https://tools.ietf.org/html/rfc8693. */
-public final class StsRequestHandler {
+final class StsRequestHandler {
   private static final String TOKEN_EXCHANGE_GRANT_TYPE =
       "urn:ietf:params:oauth:grant-type:token-exchange";
   private static final String REQUESTED_TOKEN_TYPE =
